+++ packages/cursorless-tutorial/src/index.ts
@@ -1,1 +1,2 @@
+export * from "./Tutorial";
 export * from "./TutorialImpl";
